클라이언트 측에서 base64로 생성한 이미지를 처리하는 방법을 user edit에 구현 (recipe 에서는 고려)

+++ server/src/services/user.ts
@@ -50,9 +50,11 @@ export class UserService {
     userId: IUser["_id"],
     userData: Partial<IUserUpdateInputDTO>
   ) {
-    const user = await this.readUserById(userId);
-    CloudinaryService.deleteFiles([user?.picture || ""]);
-
+    if(userData.picture) {
+      const user = await this.readUserById(userId);
+      CloudinaryService.deleteFiles([user?.picture || ""]);
+    }
+    
     return User.findByIdAndUpdate(userId, { $set: userData }, { new: true });
   }
